Add files via upload
# Project Description: Annual Sales Report for Vrinda Store

I am a data analyst, and I have created an annual sales report for Vrinda Store to help increase their sales. The report was initially created in Excel and involved several steps, including data cleaning, data processing, data analysis, and the creation of pivot tables, Data visualization dashboard, Geting to insights, Giving the solution on the basis of their Data Insights. 

The main objectives of the report were to provide insights into the annual sales, understand the customer base, and identify strategies to grow sales. Through the analysis, I discovered that targeting women between the ages of 20 and 35 in the main cities of Maharashtra, Karnataka, and Uttarakhand would be the most effective approach. Additionally, using online ads to reach this age group was recommended based on the analysis.

The report includes detailed findings and recommendations to help Vrinda Store improve their sales. By following the insights and recommendations provided in the report, Vrinda Store can strategically target their marketing efforts and increase their sales.
# Dashboard Visualization on Excel

In addition to the annual sales report, I have also created an amazing dashboard visualization on Excel. The dashboard is designed to be easy to read and provides a visual representation of the key insights and findings from the sales analysis.